[Feat] TestManager 작성 및 Presenter 수정
- 테스트를 위한 TestManager 작성
- Presenter에서 UpdateHp의 슬라이더 값 계산 시 float로 형변환하여 소수점이 나오도록 수정

+++ Assets/7. MVP/Scripts/Presenter.cs
@@ -43,7 +43,7 @@ public class Presenter : MonoBehaviour
         if (hpBar == null || model.MaxHp == 0)
             return;
 
-        hpBar.value = model.CurrentHp / model.MaxHp;
+        hpBar.value = (float)model.CurrentHp / model.MaxHp;
         txtHp.text = $"{model.CurrentHp}";
     }
